requirements.txt
langchain-huggingface
huggingface_hub
langchain
python-dotenv
streamlit
